Pending latest tests

+++ include/philo_structs.h
@@ -6,7 +6,7 @@
 /*   By: pvilchez <pvilchez@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/27 17:17:18 by pvilchez          #+#    #+#             */
-/*   Updated: 2023/12/06 19:20:28 by pvilchez         ###   ########.fr       */
+/*   Updated: 2023/12/06 21:22:25 by pvilchez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ typedef struct s_philo
 	pthread_mutex_t	*left_fork;
 	pthread_mutex_t	*right_fork;
 	pthread_mutex_t	last_eat_mutex;
-	int				last_eat;
+	long			last_eat;
 	pthread_mutex_t	num_eats_mutex;
 	int				num_eats;
 	t_table			*table;
@@ -36,10 +36,10 @@ typedef struct s_table
 	t_philo			*philos;
 	pthread_mutex_t	*forks;
 	pthread_mutex_t	print_mutex;
-	int				init_time;
-	int				time_die;
-	int				time_eat;
-	int				time_sleep;
+	long			init_time;
+	long			time_die;
+	long			time_eat;
+	long			time_sleep;
 	int				num_eats;
 	pthread_mutex_t	end_mutex;
 	int				end;
